Make several mask nodes work with batches
Updated nodes:

* WAS_Mask_Minority_Region
* WAS_Mask_Arbitrary_Region
* WAS_Mask_Fill_Region
* WAS_Mask_Threshold_Region
* WAS_Mask_Floor_Region
* WAS_Mask_Ceiling_Region
* WAS_Mask_Gaussian_Region

Not sure what the previous code was meant to check for, with the dimension
checking. It looks to me the code was working by accident, and this
understanding is reflected by my replacing the if-clause by an assert.
I may be wrong, but it didn't work for me except for singletons, and now it
works both for singletons and batches.